Move ValidateSeekableStreamAt0AndGetCount into streams sub-package
Keep the root internal package empty of functional code, grouping
related helpers into sub-packages as Joel suggested.
