+++ sdk/storage/internal/streams/streams.go
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-package internal
+package streams
 
 import (
 	"errors"
